backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.0
python-dotenv>=1.0.0
nltk>=3.8.1
textblob>=0.17.1
openai>=0.27.0
requests>=2.28.2
huggingface_hub>=0.23.0